requirements.txt
netaddr2==0.7.19
numpy==1.20.0
pandas==1.2.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
